# Use newest standard library as base for new version of ed25519 base