--- a/.gitignore
+++ b/.gitignore
@@ -13,3 +13,6 @@
 
 .DS_Store
 .idea/
+
+# file is regenerated on every testrun
+reference/testdata.csv