--- a/ed25519_ext.go
+++ b/ed25519_ext.go
@@ -26,70 +26,50 @@ func ExtractPublicKey(message, sig []byte) (PublicKey, error) {
 	// we remove the public key from the hash
 	//h.Write(privateKey[32:])
 	h.Write(message)
-	var digest [64]byte
+	digest := make([]byte, 64)
 	h.Sum(digest[:0])
 
-	var hReduced [32]byte
-	edwards25519.ScReduce(&hReduced, &digest)
-
-	var hInv [32]byte
-	edwards25519.InvertModL(&hInv, &hReduced)
-
-	var s [32]byte
-	if l := copy(s[:], sig[32:]); l != PublicKeySize {
-		return nil, errors.New("memory copy failed")
+	hReduced, err := new(edwards25519.Scalar).SetUniformBytes(digest)
+	if err != nil {
+		return nil, err
 	}
 
-	// https://tools.ietf.org/html/rfc8032#section-5.1.7 requires that s be in
-	// the range [0, order) in order to prevent signature malleability.
-	if !edwards25519.ScMinimal(&s) {
-		return nil, errors.New("invalid signature")
+	var hInv [32]byte
+	edwards25519.InvertModL(&hInv, (*[32]byte)(hReduced.Bytes()))
+
+	s, err := new(edwards25519.Scalar).SetCanonicalBytes(sig[32:])
+	if err != nil {
+		return nil, err
 	}
 
 	// var zero [32]byte
-	var one [32]byte
-	one[0] = byte(1)
+	oneBytes := make([]byte, 32)
+	oneBytes[0] = 1
+	one, err := new(edwards25519.Scalar).SetCanonicalBytes(oneBytes)
+	if err != nil {
+		return nil, err
+	}
 
 	// Extract R = sig[32:] as a point on the curve (and compute the inverse of R)
-	var R edwards25519.ExtendedGroupElement
-	var r [32]byte
-	copy(r[:], sig[:32])
-	if ok := R.FromBytes(&r); !ok {
-		return nil, errors.New("failed to create extended group element from s")
+	R, err := new(edwards25519.Point).SetBytes(sig[:32])
+	if err != nil {
+		return nil, err
 	}
 
 	// The following lines make R -> -R
-	edwards25519.FeNeg(&R.X, &R.X)
-	edwards25519.FeNeg(&R.T, &R.T)
-	var A edwards25519.ProjectiveGroupElement
-	var A2 edwards25519.ExtendedGroupElement
-	edwards25519.GeDoubleScalarMultVartime(&A, &one, &R, &s)
-	A.ToExtended(&A2)
+	R.Negate(R)
 
-	// THESE COMMENTS ARE OBSOLETE (afterwards we added ToExtended() ) -- KEPT TO REMIND OURSELVES THE LOGIC
-	// We need to convert A from projective to extended group element
-	// ToBytes takes projective
-	// FromBytes return extended
-	// Let's try....  [in general there should be a smarter way of doing this, so remember to look into this]
+	A := new(edwards25519.Point)
+	A.VarTimeDoubleScalarBaseMult(one, R, s)
 
-	// following is old commands, which we keep for benchmarking
-	//var buff [32]byte
-	//A.ToBytes(&buff)
-	//var A2 edwards25519.ExtendedGroupElement
-	//if ok := A2.FromBytes(&buff); !ok {
-	//	return nil, errors.New("failed to create an extended group element A2 from A")
-	//}
+	hInvScalar, err := new(edwards25519.Scalar).SetCanonicalBytes(hInv[:])
+	if err != nil {
+		return nil, err
+	}
 
-	var EC_PK edwards25519.ProjectiveGroupElement
-	// This is the old command, used for the built-in scalar multiplication function
-	//edwards25519.GeDoubleScalarMultVartime(&EC_PK, &hInv, &A2, &zero)
-	edwards25519.GeScalarMultVartime(&EC_PK, &hInv, &A2)
+	EC_PK := new(edwards25519.Point).ScalarMult(hInvScalar, A)
 
-	var pubKey [PublicKeySize]byte
-
-	// EC_PK is supposed to be the public key as an elliptic curve point, we apply ToBytes
-	EC_PK.ToBytes(&pubKey)
-	return pubKey[:], nil
+	return EC_PK.Bytes(), nil
 }
 
 // NewDerivedKeyFromSeed calculates a private key from a 32 bytes random seed, an integer index and salt
@@ -128,11 +108,10 @@ func Sign2(privateKey PrivateKey, message []byte) []byte {
 	// and the last 32 bytes is the (encoding) of the public key (elliptic curve point,
 	// as in line 4 in "Algorithm 1").
 	h.Write(privateKey[:32])
-
-	var digest1, messageDigest, hramDigest [64]byte
-	var expandedSecretKey [32]byte
+	digest1 := make([]byte, 64)
+	expandedSecretKey := make([]byte, 32)
 	h.Sum(digest1[:0])
-	copy(expandedSecretKey[:], digest1[:])
+	copy(expandedSecretKey, digest1)
 	expandedSecretKey[0] &= 248
 	expandedSecretKey[31] &= 63
 	expandedSecretKey[31] |= 64 // this is the final value for 'a'
@@ -145,18 +124,19 @@ func Sign2(privateKey PrivateKey, message []byte) []byte {
 	h.Write(message)
 
 	// line 5 in "Algorithm 1": creates r
+	messageDigest := make([]byte, 64)
 	h.Sum(messageDigest[:0])
 
 	// looks like reduction mod l, this is the final r
-	var messageDigestReduced [32]byte
-	edwards25519.ScReduce(&messageDigestReduced, &messageDigest)
+	messageDigestReduced, err := new(edwards25519.Scalar).SetUniformBytes(messageDigest)
+	if err != nil {
+		panic(err)
+	}
 
 	// line 6 in "Algorithm 1": creates R
-	var R edwards25519.ExtendedGroupElement
-	edwards25519.GeScalarMultBase(&R, &messageDigestReduced)
+	R := new(edwards25519.Point).ScalarBaseMult(messageDigestReduced)
 
-	var encodedR [32]byte
-	R.ToBytes(&encodedR)
+	encodedR := R.Bytes()
 
 	h.Reset()
 	h.Write(encodedR[:])
@@ -165,21 +145,26 @@ func Sign2(privateKey PrivateKey, message []byte) []byte {
 
 	// line 7: creates h
 	h.Write(message)
+	hramDigest := make([]byte, 64)
 	h.Sum(hramDigest[:0])
 
-	var hramDigestReduced [32]byte
-
 	// this is the final h
-	edwards25519.ScReduce(&hramDigestReduced, &hramDigest)
+	hramDigestReduced, err := new(edwards25519.Scalar).SetUniformBytes(hramDigest)
+	if err != nil {
+		panic(err)
+	}
+
+	priv, err := new(edwards25519.Scalar).SetBytesWithClamping(expandedSecretKey)
+	if err != nil {
+		panic(err)
+	}
 
 	// line 8: s = h*a + r
-	var s [32]byte
-	edwards25519.ScMulAdd(&s, &hramDigestReduced, &expandedSecretKey, &messageDigestReduced)
+	s := new(edwards25519.Scalar).MultiplyAdd(hramDigestReduced, priv, messageDigestReduced)
 
 	signature := make([]byte, SignatureSize)
-	copy(signature[:], encodedR[:])
-	copy(signature[32:], s[:])
-
+	copy(signature, encodedR)
+	copy(signature[32:], s.Bytes())
 	return signature
 }
 
@@ -194,39 +179,33 @@ func Verify2(publicKey PublicKey, message, sig []byte) bool {
 		return false
 	}
 
-	var A edwards25519.ExtendedGroupElement
-	var publicKeyBytes [32]byte
-	copy(publicKeyBytes[:], publicKey)
-	if !A.FromBytes(&publicKeyBytes) {
+	A, err := new(edwards25519.Point).SetBytes(publicKey)
+	if err != nil {
 		return false
 	}
-	edwards25519.FeNeg(&A.X, &A.X)
-	edwards25519.FeNeg(&A.T, &A.T)
+	A.Negate(A)
 
 	h := sha512.New()
 	h.Write(sig[:32])
 	// we remove the public key from the hash
 	// h.Write(publicKey[:])
 	h.Write(message)
-	var digest [64]byte
+	digest := make([]byte, 64)
 	h.Sum(digest[:0])
 
-	var hReduced [32]byte
-	edwards25519.ScReduce(&hReduced, &digest)
-
-	var R edwards25519.ProjectiveGroupElement
-	var s [32]byte
-	copy(s[:], sig[32:])
-
-	// https://tools.ietf.org/html/rfc8032#section-5.1.7 requires that s be in
-	// the range [0, order) in order to prevent signature malleability.
-	if !edwards25519.ScMinimal(&s) {
+	hReduced, err := new(edwards25519.Scalar).SetUniformBytes(digest)
+	if err != nil {
 		return false
 	}
 
-	edwards25519.GeDoubleScalarMultVartime(&R, &hReduced, &A, &s)
+	s := make([]byte, 32)
+	copy(s, sig[32:])
 
-	var checkR [32]byte
-	R.ToBytes(&checkR)
-	return bytes.Equal(sig[:32], checkR[:])
+	priv, err := new(edwards25519.Scalar).SetCanonicalBytes(s)
+	if err != nil {
+		panic(err)
+	}
+
+	R := new(edwards25519.Point).VarTimeDoubleScalarBaseMult(hReduced, A, priv)
+	return bytes.Equal(sig[:32], R.Bytes())
 }