--- a/ed25519_ext.go
+++ b/ed25519_ext.go
@@ -16,6 +16,8 @@ import (
 // Note that signature must be created using Sign2() and NOT using Sign().
 // It will panic if len(sig) is not SignatureSize.
 func ExtractPublicKey(message, sig []byte) (PublicKey, error) {
+	// Outline the function body so that the returned PublicKey can be
+	// stack-allocated.
 	publicKey := make([]byte, PublicKeySize)
 	err := extractPublicKey(publicKey, message, sig)
 	return publicKey, err