--- a/go.mod
+++ b/go.mod
@@ -2,7 +2,7 @@ module github.com/spacemeshos/ed25519
 
 go 1.19
 
-require github.com/stretchr/testify v1.8.1
+require github.com/stretchr/testify v1.8.2
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect