--- a/internal/edwards25519/field/_asm/go.mod
+++ b/internal/edwards25519/field/_asm/go.mod
@@ -5,7 +5,7 @@ go 1.19
 require github.com/mmcloughlin/avo v0.5.0
 
 require (
-	golang.org/x/mod v0.6.0 // indirect
-	golang.org/x/sys v0.1.0 // indirect
-	golang.org/x/tools v0.2.0 // indirect
+	golang.org/x/mod v0.8.0 // indirect
+	golang.org/x/sys v0.5.0 // indirect
+	golang.org/x/tools v0.5.0 // indirect
 )