… .github/workflows/ci.yml
… ed25519_ext_test.go
… ed25519_test.go
… go.sum
… internal/edwards25519/const.go
… internal/edwards25519/edwards25519.go
… internal/edwards25519/edwards25519_test.go
… internal/edwards25519/field/_asm/fe_amd64_asm.go
… internal/edwards25519/field/_asm/go.sum
… internal/edwards25519/field/fe.go
… internal/edwards25519/field/fe_alias_test.go
… internal/edwards25519/field/fe_amd64.go
… internal/edwards25519/field/fe_amd64.s
… internal/edwards25519/field/fe_bench_test.go
… internal/edwards25519/field/fe_generic.go
… internal/edwards25519/field/fe_test.go
… internal/edwards25519/invert.go
… internal/edwards25519/invert_test.go
… internal/edwards25519/scalar.go
… internal/edwards25519/scalar_alias_test.go
… internal/edwards25519/scalar_fiat.go
… internal/edwards25519/scalar_test.go
… internal/edwards25519/scalarmult.go
… internal/edwards25519/scalarmult_test.go
… internal/edwards25519/tables_test.go
… reference/requirements.txt
… testdata/sign.input.gz