This improves the extensions for ED25519 to be more efficient. They are now on par with the ed25519 versions of the standard library:

without upgrade:

```
BenchmarkPublicKeyExtraction-6              6681            174333 ns/op              32 B/op          1 allocs/op
BenchmarkSigningExt-6                      35907             33093 ns/op              64 B/op          1 allocs/op
BenchmarkVerificationExt-6                 13371             89331 ns/op               0 B/op          0 allocs/op
BenchmarkSigning-6                         54087             22482 ns/op               0 B/op          0 allocs/op
BenchmarkVerification-6                    25714             46642 ns/op               0 B/op          0 allocs/op
```

with this upgrade:

```
BenchmarkPublicKeyExtraction-6             10000            109794 ns/op              32 B/op          1 allocs/op
BenchmarkSigningExt-6                      65743             18310 ns/op              64 B/op          1 allocs/op
BenchmarkVerificationExt-6                 25792             46711 ns/op               0 B/op          0 allocs/op
BenchmarkSigning-6                         54050             22240 ns/op               0 B/op          0 allocs/op
BenchmarkVerification-6                    25460             46372 ns/op               0 B/op          0 allocs/op
```

`SigningExt` and `VerificationExt` are now on par (or even faster) than their standard library versions and `ExtractPublicKey` has been improved in speed by ~ 40%.