@pigmej 2023-02-23
```
[poet@nyc-poet ed25519]$ ~/go/bin/go test -benchmem -bench=. github.com/spacemeshos/ed25519
goos: linux
goarch: amd64
pkg: github.com/spacemeshos/ed25519
cpu: 13th Gen Intel(R) Core(TM) i9-13900KS
BenchmarkPublicKeyExtraction-17    	   10000	    116848 ns/op	      32 B/op	       1 allocs/op
BenchmarkSigningExt-17             	   57896	     19610 ns/op	      64 B/op	       1 allocs/op
BenchmarkVerificationExt-17        	   19662	     59374 ns/op	       0 B/op	       0 allocs/op
BenchmarkSigning-17                	  110148	     11133 ns/op	       0 B/op	       0 allocs/op
BenchmarkVerification-17           	   43221	     24328 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/spacemeshos/ed25519	7.857s
[poet@nyc-poet ed25519]$ git checkout
derive                                 origin/develop                         upgrade-code-to-latest-stdlib
develop                                origin/HEAD                            v0.1.0
HEAD                                   origin/master                          v0.1.1
master                                 origin/upgrade-code-to-latest-stdlib   vrf
origin/derive                          origin/vrf
[poet@nyc-poet ed25519]$ git checkout upgrade-code-to-latest-stdlib
branch 'upgrade-code-to-latest-stdlib' set up to track 'origin/upgrade-code-to-latest-stdlib'.
Switched to a new branch 'upgrade-code-to-latest-stdlib'
[poet@nyc-poet ed25519]$ ~/go/bin/go test -benchmem -bench=. github.com/spacemeshos/ed25519
goos: linux
goarch: amd64
pkg: github.com/spacemeshos/ed25519
cpu: 13th Gen Intel(R) Core(TM) i9-13900KS
BenchmarkPublicKeyExtraction-17    	   20401	     65809 ns/op	      32 B/op	       1 allocs/op
BenchmarkSigningExt-17             	  110592	      9514 ns/op	      64 B/op	       1 allocs/op
BenchmarkVerificationExt-17        	   49764	     27052 ns/op	       0 B/op	       0 allocs/op
BenchmarkSigning-17                	   91276	     11410 ns/op	       0 B/op	       0 allocs/op
BenchmarkVerification-17           	   47253	     24732 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/spacemeshos/ed25519	8.126s
```

Works on AMD64 too.