@fasmat 2023-02-23
I did some additional improvements. Now Sign2 and Verify2 are basically the same as the standard library versions (also almost exactly the same code) and `ExtractPublicKey` is now 2.5x faster than the previous implementation:

```
goos: linux
goarch: arm64
pkg: github.com/spacemeshos/ed25519
BenchmarkPublicKeyExtraction-6             17200             69598 ns/op               0 B/op          0 allocs/op
BenchmarkSigningExt-6                      53544             22400 ns/op               0 B/op          0 allocs/op
BenchmarkVerificationExt-6                 25791             46504 ns/op               0 B/op          0 allocs/op
BenchmarkSigning-6                         54090             22515 ns/op               0 B/op          0 allocs/op
BenchmarkVerification-6                    25606             46799 ns/op               0 B/op          0 allocs/op
PASS
ok      github.com/spacemeshos/ed25519  8.130s
```

@pigmej can you want to try again on AMD64? I assume the latest version will be even faster than the last version you tested 🙂